backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.0
django-filter==23.5
google-auth==2.29.0
requests==2.31.0
mercadopago==2.2.1